修复了大数字插入失败的问题，将国创爬取顺序改为了播放量

+++ anime/anime/initialMySQL.py
@@ -17,7 +17,7 @@ def pymysql_create_table():
             name text not null,
             score float not null,
             score_num int not null,
-            play int not null,
+            play bigint not null,
             follow int not null,
             barrage int not null,
             tags text not null,
